requirements.txt
pyrogram
python-decouple